Create RE👋 Hi, I’m @randelllmalavidaADME.md
I am very happy to be in this community of very talented individuals, for whom my admiration and respect began, when I got my first Apple/Macintosh floppy driven, second hand personal computer. The future was now and I had arrived. 
I was so ready to lose myself into this new cyber world locked inside this small powerhouse  of infinite technological possibilities. I said goodbye to a few friends and told my family I'd be away for a while. As my mind conjured up what my virtual reality world would be like, I saw myself floating over pyramids and then walking aboard the Love Boat, enroute to charter a small plane, I would pilot myself and land onto fantasy island. "Boss the plane, the plane!!!"... 
Hold thought!!! Here comes Geo with my device, wrapped in a brown paper grocery bag. "Awesome bro!!!", how much???!!!, Ah man, that means Chapter 13~ So, after getting rung up , I dimmed the lights, plugged it in and finally the moment I had imagined. I hit the power and the floppy disk engaged and the room came alive and sounded like a busy news room.  The screen came on, an icon blinked as the floppy was spindling, spitting and working on something~ for a very long time. It had to work after hours of trying to move the mouse to fix it, I decided to gingerly pick up this modern marvel of human achievement and placed the machine w/ flashing icon on the screen to my forehead and began chanting, "come to me, come to me,  I call the spirit of this spiders web~ come to me....". The screen flashed and all the light appeared to get sucked into a single point and went black and the machine fell silent into the darkness of my room. I stayed in there for days, "maybe I'm in the web now??? was this it???  Because, if it is~ it sucks!!! How are these kids able to Travers this with such ease??? This was in 2003.  I'm still new to this aspect of using these tools, however should my skill set improve, I would love to create something useful, fun that may help our community and our world. 
My appreciation for the ability of computer savvy people, or those that want to learn as in this Community, I have great respect and hold in high regard. 
Thank You!!!
Randy Chavez.   kickrocks@toothfairy.com

I graduated from the San Francisco Art Institute with a BFA. Studied Percussion and Religious Ceremonies in Afro Brazilian Culture and would you believe while doing an internship at the Hall of Justice, Inspector's Bureau of the San Francisco Police Department, one of my duties was to input the data from hard copy into their database using the same dinosaur Macintosh box.